added servo-test
12/4/20
